Première vague marave OK

+++ www/superadmin/do/doDeleteChamp.php
@@ -2,33 +2,44 @@
 session_start();
 
 if ($_SESSION["superadmin"]) {
-    include("../../includes/mysqli.php");
+    include("../../includes/PDO.php");
     include("../../includes/status.php");
     
-    $query = "DELETE FROM `champ` 
-            WHERE `pk_champ` = " . $_POST["pk"] . " 
-                AND  `fk_client` = " . $_POST["client"] . " 
-                AND `fk_monde` = " . $_POST["monde"] . "; 
-            DELETE FROM `valeur_champ` 
-            WHERE `fk_champ` = " . $_POST["pk"] . "
-                AND  `fk_client` = " . $_POST["client"] . " 
-                AND `fk_monde` = " . $_POST["monde"] . ";";
+    $query = "
+        DELETE FROM `champ` 
+        WHERE `pk_champ` = :pk
+            AND  `fk_client` = :client 
+            AND `fk_monde` = :monde; "
+            
+    $result = dino_query($query,[
+        "client" => $_POST["client"],
+        "monde" => $_POST["monde"],
+        "pk" => $_POST["pk"]
+    ]);
     
-    if ($mysqli->multi_query($query)) {
-        $i = 0; 
-        do { 
-            $i++; 
-        } while ($mysqli->next_result()); 
+    if ($result["status"]) {
+        $query = "
+            DELETE FROM `valeur_champ` 
+            WHERE `fk_champ` = :pk
+                AND  `fk_client` = :client 
+                AND `fk_monde` = :monde
+        ;";     
         
-        if (!$mysqli->errno) { 
+        $result = dino_query($query,[
+            "client" => $_POST["client"],
+            "monde" => $_POST["monde"],
+            "pk" => $_POST["pk"]
+        ]);
+        
+        if ($result["status"]) { 
             status(200);
         } else {
             status(500);
-            $json = '{ "error": "mysqli", "query": "' . $query . '", "message": "' . $mysqli->error . '" }';
+            $json = '{ "error": "mysqli", "query": "' . $query . '", "message": "' . $result["errinfo"][2] . '" }';
         }
     } else {
         status(500);
-        $json = '{ "error": "mysqli", "query": "' . $query . '", "message": "' . $mysqli->error . '" }';
+        $json = '{ "error": "mysqli", "query": "' . $query . '", "message": "' . $result["errinfo"][2] . '" }';
     }
     
     header('Content-Type: application/json');
